refactor: #9 ListAdapter 관련 파일 정리
